small changes game engine, fixed save gui position
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2027 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/kingdom/kingdom.cpp
@@ -618,7 +618,7 @@ const Heroes* Kingdom::GetBestHero(void) const
 
 u32 Kingdom::GetArmiesStrength(void) const
 {
-    u32 res = 0;
+    double res = 0;
 
     std::vector<Heroes*>::const_iterator ith = heroes.begin();
     for(; ith != heroes.end(); ++ith) if(*ith) res += (**ith).GetArmy().GetStrength();
